allow recursion only for our net
We do not want to be a nameserver for everyone knowing us, we only want to serve to our net. But everyone is allowed to ask stuff about our net (like „who is draco.fftr“) and we will happily answer. Other stuff (like „who is google.de“) will only get the „how should I know“-answer.