Reduced TTL from 2 days to 2 hours
I just noticed that the routers wouldn't update the firmware. I guess that they have the old ip cached and will wait up to 2 days to request another update. I think 2 hours is a more reasonable value because we have the DNS to be able to change things easyly and that should not take that many ages